tensorflow/requirements.txt
tensorflow>=2.0.0
tensorflow-hub
tensorflow-datasets
numpy
matplotlib
pandas
pillow
